Crate cu29_log_runtime

Structs

LoggerRuntime

The lifetime of this struct is the lifetime of the logger.

NullLog

OwningIoWriter

SimpleFileWriter

This allows this crate to be used outside of Copper (ie. decoupling it from the unifiedlog.

Statics

EXTRA_TEXT_LOGGER

Functions

log

Function called from generated code to log data. It moves entry by design, it will be absorbed in the queue.

log_debug_mode

This version of log is only compiled in debug mode This allows a normal logging framework to be bridged.

structured_log_bytes_total

Returns the total number of bytes written to the structured log stream.